add base_priority_rule.py for setting dispaching rule of each task
