+++ pDESy/model/base_priority_rule.py
@@ -1,0 +1,16 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from enum import IntEnum
+
+
+class ResourcePriorityRule(IntEnum):
+    """ResourcePriorityRule"""
+
+    SSP = 0
+
+
+class TaskPriorityRule(IntEnum):
+    """TaskPriorityRule"""
+
+    TSLACK = 0
