+++ pDESy/model/base_task.py
@@ -5,6 +5,7 @@ import abc
 import uuid
 from enum import IntEnum
 import datetime
+from .base_priority_rule import ResourcePriorityRule
 
 
 class BaseTaskState(IntEnum):
@@ -59,6 +60,12 @@ class BaseTask(object, metaclass=abc.ABCMeta):
             Basic parameter.
             List of allocated BaseFactory
             Defaults to None -> [].
+        worker_priority_rule (ResourcePriorityRule, oprional):
+            Worker priority rule for simulation.
+            Deraults to ResourcePriorityRule.SSP.
+        facility_priority_rule (ResourcePriorityRule, oprional):
+            Task priority rule for simulation.
+            Deraults to TaskPriorityRule.TSLACK.
         need_facility (bool, optional):
             Basic parameter.
             Whether one facility is needed for performing this task or not.
@@ -143,6 +150,8 @@ class BaseTask(object, metaclass=abc.ABCMeta):
         output_task_list=None,
         allocated_team_list=None,
         allocated_factory_list=None,
+        worker_priority_rule=ResourcePriorityRule.SSP,
+        facility_priority_rule=ResourcePriorityRule.SSP,
         need_facility=False,
         target_component=None,
         default_progress=None,
@@ -180,6 +189,16 @@ class BaseTask(object, metaclass=abc.ABCMeta):
         )
         self.allocated_factory_list = (
             allocated_factory_list if allocated_factory_list is not None else []
+        )
+        self.worker_priority_rule = (
+            worker_priority_rule
+            if worker_priority_rule is not None
+            else ResourcePriorityRule.SSP
+        )
+        self.facility_priority_rule = (
+            facility_priority_rule
+            if facility_priority_rule is not None
+            else ResourcePriorityRule.SSP
         )
         self.need_facility = need_facility
         self.target_component = (
